Created BGMTrack class and added some temp clips
Each BGMTrack object stores the AudioClips to play back, information about the clips being played, and an algorithm for deciding the order to play the clips in (currently hardcoded to play through clips in order).

I'll refactor MusicPlayer next to take BGMTrack objects as a parameter to allow drag-and-drop track switching in scenes.