Intial Player Collision & Pick-Up Interaction
Player is now constrained within the bounds of the camera's screenspace. Added tags to some things, and made objects that the player can bump into for +lives, -lives, and +points. Added some basic score and lives UI elements, and updated the music to vince's new stage 1 music in the player test scene.