Fixed player Bullet bug + Added Scrolling Background Prefab
With the aid of chris a bug was fixed to make it so that the player's bullet is now properly above all parts of the scrolling background. Also added a premade background scroller, just drop it in a scene and then drop a background, midground, and foreground into the texture for each layer.